More map creation work
Map areas are now created on the server with the proper area types.
Integration test added to check that the minimum number of areas
(according to the defaults) are created.

On second thought, this could actually fail. Need to make the test
better and take into account the modifications from the names.